render api refactor (#561)
* fix wrong choroplethmap when png size is small

* fix wrong choroplethmap when png size is small

* add test for transform_projection

* format projection test

* package python transform and render

* pylint && code format

* python support chunkarray

* choroplethmap without agg, done

* chroplethmap with agg, done

* icion map, done

* heat map, done

* point map, done

* modify weighted pointmap in c++

* wrap weighted pointmap in cython

* wrap weighted pointmap in pandas api, done

* projection api, done

* fix transform api

* use flexible data type

* render code refactor, done

* remove redundant print msg

* code format

* use _to_arrow_array_list and _to_pandas_series

* remove render raw point arrow api

* add agg handler

* code format

* code format
