+++ cpp/src/arrow/render_api.h
@@ -37,10 +37,6 @@ const std::vector<std::shared_ptr<arrow::Array>> transform_and_projection(
     const std::string& dst_rs, const std::string& bottom_right,
     const std::string& top_left, const int& height, const int& width);
 
-std::shared_ptr<arrow::Array> point_map(const std::shared_ptr<arrow::Array>& arr_x,
-                                        const std::shared_ptr<arrow::Array>& arr_y,
-                                        const std::string& conf);
-
 std::shared_ptr<arrow::Array> point_map(
     const std::vector<std::shared_ptr<arrow::Array>>& points_vector,
     const std::string& conf);
@@ -50,31 +46,18 @@ std::shared_ptr<arrow::Array> weighted_point_map(
     const std::vector<std::shared_ptr<arrow::Array>>& points_vector,
     const std::string& conf);
 
-// three args api: point_map(x, y, conf), point_map(wkt, c, conf), point_map(wkt, s, conf)
+// three args api: point_map(wkt, c, conf), point_map(wkt, s, conf)
 std::shared_ptr<arrow::Array> weighted_point_map(
     const std::vector<std::shared_ptr<arrow::Array>>& points_vector,
     const std::vector<std::shared_ptr<arrow::Array>>& weights_vector,
     const std::string& conf);
 
-// four args api: point_map(x, y, c, conf), point_map(x, y, s, conf), point_map(wkt, c, s,
-// conf)
+// four args api: point_map(wkt, c, s, conf)
 std::shared_ptr<arrow::Array> weighted_point_map(
     const std::vector<std::shared_ptr<arrow::Array>>& points_vector,
     const std::vector<std::shared_ptr<arrow::Array>>& color_weights_vector,
     const std::vector<std::shared_ptr<arrow::Array>>& size_weights_vector,
     const std::string& conf);
-
-// five args api: point_map(x, y, c, s, conf)
-std::shared_ptr<arrow::Array> weighted_point_map(
-    const std::shared_ptr<arrow::Array>& arr_x,
-    const std::shared_ptr<arrow::Array>& arr_y,
-    const std::shared_ptr<arrow::Array>& arr_c,
-    const std::shared_ptr<arrow::Array>& arr_s, const std::string& conf);
-
-std::shared_ptr<arrow::Array> heat_map(const std::shared_ptr<arrow::Array>& arr_x,
-                                       const std::shared_ptr<arrow::Array>& arr_y,
-                                       const std::shared_ptr<arrow::Array>& arr_c,
-                                       const std::string& conf);
 
 std::shared_ptr<arrow::Array> heat_map(
     const std::vector<std::shared_ptr<arrow::Array>>& points_vector,
@@ -85,10 +68,6 @@ std::shared_ptr<arrow::Array> choropleth_map(
     const std::vector<std::shared_ptr<arrow::Array>>& polygons_vector,
     const std::vector<std::shared_ptr<arrow::Array>>& weights_vector,
     const std::string& conf);
-
-std::shared_ptr<arrow::Array> icon_viz(const std::shared_ptr<arrow::Array>& arr_x,
-                                       const std::shared_ptr<arrow::Array>& arr_y,
-                                       const std::string& conf);
 
 std::shared_ptr<arrow::Array> icon_viz(
     const std::vector<std::shared_ptr<arrow::Array>>& points_vector,
