+++ cpp/src/render/utils/agg/agg_handler.cpp
@@ -1,0 +1,33 @@
+/*
+ * Copyright (C) 2019-2020 Zilliz. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+#include "render/utils/agg/agg_handler.h"
+
+namespace arctern {
+namespace render {
+
+AggHandler::AggType AggHandler::agg_type(std::string type) {
+  if (type == "mean") return AggType::AVG;
+  if (type == "sum") return AggType::SUM;
+  if (type == "max") return AggType::MAX;
+  if (type == "min") return AggType::MIN;
+  if (type == "count") return AggType::COUNT;
+  if (type == "std") return AggType::STDDEV;
+  std::string err_msg = "unknow agg type = " + type;
+  throw std::runtime_error(err_msg);
+}
+
+}  // namespace render
+}  // namespace arctern
